(fixed) change funtion definition from data config to dataset
