+++ matorage/data/config.py
@@ -334,44 +334,3 @@ class DataConfig(StorageConfig):
             return 0
         else:
             return keys[-1]
-
-    @property
-    def get_filetype_list(self):
-        """
-        Get list of filetype dataset in bucket of ``DataConfig``
-
-        Returns:
-            :obj: `list`: list of key of filetype dataset in bucket of ``DataConfig``
-        """
-        return list(self.metadata.filetype.keys())
-
-    def get_filetype_from_key(self, key):
-        """
-        Download filetype dataset from key in bucket of ``DataConfig``
-
-        Args:
-            key (:obj:`string`):
-                key name of file which you will download.
-
-        Returns:
-            :obj: `string`: local path of downloaded file in bucket of ``DataConfig``
-        """
-        _client = (
-            Minio(
-                self.endpoint,
-                access_key=self.access_key,
-                secret_key=self.secret_key,
-                secure=self.secure,
-                region=self.region,
-            )
-            if not check_nas(self.endpoint)
-            else NAS(self.endpoint)
-        )
-
-        _local_file = tempfile.mktemp(key)
-        _client.fget_object(
-            bucket_name=self.bucket_name,
-            object_name=key,
-            file_path=_local_file
-        )
-        return _local_file
